+++ src/main/java/Arquitetura/Model/Medico.java
@@ -27,14 +27,14 @@ public class Medico extends Usuario {
     }
 
     // Não possui subEspecialidade e possui ID
-    public Medico(int id, String nome, String cpf, String senha, double salario, int cargaHorariaSemanal, String plantao, String especialidade, String formacao, String sexo, String telefone, String email, Date dataNascimento) {
+    public Medico(long id, String nome, String cpf, String senha, double salario, int cargaHorariaSemanal, String plantao, String especialidade, String formacao, String sexo, String telefone, String email, Date dataNascimento) {
         this(nome, cpf, senha, salario, cargaHorariaSemanal, plantao, especialidade, formacao, sexo, telefone, email, dataNascimento);
         this.subEspecialidade = "N/A";
         this.setId(id);
     }
 
     // Possui subEspecialidade e possui ID
-    public Medico(int id, String nome, String cpf, String senha, double salario, int cargaHorariaSemanal, String plantao, String especialidade, String formacao, String subEspecialidade, String sexo, String telefone, String email, Date dataNascimento) {
+    public Medico(long id, String nome, String cpf, String senha, double salario, int cargaHorariaSemanal, String plantao, String especialidade, String formacao, String subEspecialidade, String sexo, String telefone, String email, Date dataNascimento) {
         this(nome, cpf, senha, salario, cargaHorariaSemanal, plantao, especialidade, formacao, sexo, telefone, email, dataNascimento);
         this.subEspecialidade = subEspecialidade;
         this.setId(id);
